update: スタートの安定化

+++ etrobo_tr/include/BalancerCpp.h
@@ -20,7 +20,7 @@ class Balancer
 public:
     Balancer();
 
-    void init(int offset);
+    void init(float offset);
     void update(int angle, int rwEnc, int lwEnc, int battery);
     void setCommand(int forward, int turn);
     int8_t getPwmRight();
